Revert "merge fix"
This reverts commit 75306ae24aec8bd8ae7833ae3915142cfcc895bb, reversing
changes made to 31a04d992c7cddbba742416889d4ccd1745bbc44.

+++ espnet/nets/pytorch_backend/nets_utils.py
@@ -151,8 +151,7 @@ def make_pad_mask(lengths, xs=None, length_dim=-1, maxlen=None):
         raise ValueError("length_dim cannot be 0: {}".format(length_dim))
 
     if not isinstance(lengths, list):
-        lengths = lengths.long().tolist()
-
+        lengths = lengths.tolist()
     bs = int(len(lengths))
     if maxlen is None:
         if xs is None:
